backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
sqlalchemy==2.0.44
PyMySQL==1.1.2
PyJWT==2.10.1
langchain==0.3.27
langchain-core==0.3.76
langchain-deepseek==0.1.4
httpx==0.28.1
python-json-logger==3.3.0
pytest==8.4.1
pytest-asyncio==1.1.0
